[INPLACE-169] 지도 관리 및 드래그 함수 훅으로 분리
